Update karma configuration to use system js for loading tests
We are migrating all of our UI packages to use SystemJS for loading and bundling. We also want to migrate the tests for several reasons. First, we are not straddling two technologies this way (Webpack and SystemJS). Second, Webpack consistantly had a problem with Angular 2 where multiple instances were being loaded. Using SystemJS resolves this issue.